+++ packages/astro/src/core/add/consts.ts
@@ -18,7 +18,7 @@ export const ALIASES = new Map([
 ]);
 export const CONFIG_STUB = `import { defineConfig } from 'astro/config';\n\nexport default defineConfig({});`;
 export const TAILWIND_CONFIG_STUB = `module.exports = {
-	content: [],
+	content: ['./src/**/*.{astro,html,js,jsx,svelte,ts,tsx,vue}'],
 	theme: {
 		extend: {},
 	},